Renamed armory item select
