--- src/app/shared/dialogs/AddArmoryItemDialog/ArmoryItemSelect.css
@@ -1,19 +1,0 @@
-.armory-item-select-card {
-	height: 220px;
-	width: 30%;
-	min-width: 0;
-	margin-right: 0;
-}
-
-.loadout-select-list {
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: space-evenly;
-}
-
-@media (max-width: 768px) {	
-	.armory-item-select-card {
-		height: 170px;
-		width: 45%;
-	}
-}
