--- src/app/shared/dialogs/AddArmoryItemDialog/index.js
@@ -1,2 +1,0 @@
-import AddArmoryItemDialog from './AddArmoryItemDialog'
-export default AddArmoryItemDialog
