+++ src/app/shared/dialogs/AddResourceDialog/index.js
@@ -1,0 +1,2 @@
+import AddResourceDialog from './AddResourceDialog'
+export default AddResourceDialog
