+++ src/app/shared/resources/ResourceList.js
