+++ src/app/shared/resources/ResourceListContainer.js
