+++ src/app/shared/resources/index.js
